+++ scripts/graphify_smoke/__init__.py
@@ -1,0 +1,1 @@
+"""graphify Docker graph/report smoke modules."""
