+++ scripts/graphify_smoke/benchmark.py
@@ -1,0 +1,142 @@
+from __future__ import annotations
+
+import json
+import subprocess
+from dataclasses import replace
+from pathlib import Path
+from typing import Any
+
+from .common import rel
+from .context import REPORT_JSON, REPORT_MD, ROOT_DIR
+from .models import StatusState
+
+
+
+def run_scored_report(fixture_path: Path, manifest_path: Path, status: StatusState) -> dict[str, Any]:
+    """Score the generated graphify fixture through the real-world job runner."""
+
+    run_cmd = [
+        "cargo",
+        "run",
+        "-p",
+        "elf-eval",
+        "--bin",
+        "real_world_job_benchmark",
+        "--",
+        "run",
+        "--fixtures",
+        str(fixture_path),
+        "--out",
+        str(REPORT_JSON),
+        "--run-id",
+        "real-world-memory-live-graphify",
+        "--adapter-id",
+        "graphify_docker_smoke",
+        "--adapter-name",
+        "graphify Docker graph/report smoke adapter",
+        "--adapter-behavior",
+        "docker_cli_graph_report_smoke",
+        "--adapter-storage-status",
+        status.setup,
+        "--adapter-runtime-status",
+        status.overall,
+        "--adapter-notes",
+        "Generated by the graphify Docker graph/report smoke; pass or wrong_result requires graph.json, GRAPH_REPORT.md, and query output mapped to generated evidence ids, while setup/runtime limits remain typed.",
+        "--external-adapter-manifest",
+        str(manifest_path),
+    ]
+    publish_cmd = [
+        "cargo",
+        "run",
+        "-p",
+        "elf-eval",
+        "--bin",
+        "real_world_job_benchmark",
+        "--",
+        "publish",
+        "--report",
+        str(REPORT_JSON),
+        "--out",
+        str(REPORT_MD),
+    ]
+
+    subprocess.run(run_cmd, cwd=ROOT_DIR, check=True)
+    subprocess.run(publish_cmd, cwd=ROOT_DIR, check=True)
+
+    report = json.loads(REPORT_JSON.read_text(encoding="utf-8"))
+
+    return {
+        "json": rel(REPORT_JSON),
+        "markdown": rel(REPORT_MD),
+        "summary": report.get("summary", {}),
+        "suites": report.get("suites", []),
+    }
+
+
+def scored_benchmark(report: dict[str, Any] | None) -> dict[str, Any]:
+    """Extract the post-score benchmark status from a real_world_job report."""
+
+    if report is None:
+        return {
+            "schema": "elf.scored_benchmark_status/v1",
+            "source": "real_world_job_benchmark",
+            "status": "pending",
+            "reason": "The smoke materialization was written before benchmark scoring completed.",
+        }
+
+    summary = report.get("summary", {})
+    counts = {
+        status: int(summary.get(status, 0) or 0)
+        for status in (
+            "pass",
+            "wrong_result",
+            "lifecycle_fail",
+            "incomplete",
+            "blocked",
+            "not_encoded",
+        )
+    }
+    status = next((name for name, count in counts.items() if name != "pass" and count > 0), "pass")
+
+    return {
+        "schema": "elf.scored_benchmark_status/v1",
+        "source": "real_world_job_benchmark",
+        "status": status,
+        "counts": counts,
+        "job_count": int(summary.get("job_count", 0) or 0),
+        "mean_score": summary.get("mean_score"),
+        "evidence_coverage": summary.get("evidence_coverage"),
+    }
+
+
+def status_with_scored_result(status: StatusState, report: dict[str, Any]) -> StatusState:
+    """Return a manifest status that follows the scored real_world_job outcome."""
+
+    scored = scored_benchmark(report)
+    scored_status = scored.get("status")
+    if scored_status not in {
+        "pass",
+        "wrong_result",
+        "lifecycle_fail",
+        "incomplete",
+        "blocked",
+        "not_encoded",
+    }:
+        return status
+
+    manifest_status = replace(status)
+    manifest_status.result = str(scored_status)
+    manifest_status.overall = str(scored_status)
+
+    if scored_status == "pass":
+        manifest_status.failure_class = ""
+        manifest_status.failure_reason = ""
+    elif scored_status == "wrong_result":
+        manifest_status.failure_class = "scored_benchmark_wrong_result"
+        manifest_status.failure_reason = (
+            "The graphify smoke materialized graph/report evidence, but the scored "
+            "real_world_job outcome is wrong_result; inspect graphify-report.json for "
+            "wrong-result signals."
+        )
+
+    return manifest_status
